+++ src/main/java/com/roboter5123/backend/config/WebSocketConfig.java
@@ -1,5 +1,4 @@
 package com.roboter5123.backend.config;
-
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
