+++ src/main/java/com/roboter5123/backend/game/implementation/chat/ChatGame.java
@@ -2,7 +2,7 @@ package com.roboter5123.backend.game.implementation.chat;
 import com.roboter5123.backend.game.api.Command;
 import com.roboter5123.backend.game.api.Game;
 import com.roboter5123.backend.game.api.GameMode;
-import com.roboter5123.backend.game.api.MessageBroadcaster;
+import com.roboter5123.backend.messaging.api.MessageBroadcaster;
 import lombok.Getter;
 import lombok.Setter;
