+++ src/main/java/com/roboter5123/backend/messaging/api/MessageBroadcaster.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.game.api;
+package com.roboter5123.backend.messaging.api;
 public interface MessageBroadcaster {
 
     void broadcastGameUpdate(String sessionCode, Object payload);
