--- src/main/java/com/roboter5123/backend/messaging/api/StompController.java
@@ -1,8 +1,0 @@
-package com.roboter5123.backend.messaging.api;
-import com.roboter5123.backend.service.model.StompMessage;
-import com.roboter5123.backend.service.model.exception.NoSuchSessionException;
-
-public interface StompController {
-
-    StompMessage joinSession(String playerName, String sessionCode) throws NoSuchSessionException;
-}
