+++ src/main/java/com/roboter5123/backend/messaging/model/JoinPayloads.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.service.model;
+package com.roboter5123.backend.messaging.model;
 import com.roboter5123.backend.game.api.Command;
 import lombok.AllArgsConstructor;
 import lombok.Data;
