+++ src/main/java/com/roboter5123/backend/messaging/model/ServiceCommand.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.service.model;
+package com.roboter5123.backend.messaging.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
