+++ src/main/java/com/roboter5123/backend/messaging/model/ServiceCommandStompMessage.java
@@ -1,16 +1,14 @@
-package com.roboter5123.backend.service.model;
+package com.roboter5123.backend.messaging.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.util.Map;
-
 @Data
 @EqualsAndHashCode
 @AllArgsConstructor
 @NoArgsConstructor
-public class StateStompMessage implements StompMessage {
+public class ServiceCommandStompMessage implements StompMessage {
 
-    Map<String, Object> state;
+    ServiceCommand command;
 }
