+++ src/main/java/com/roboter5123/backend/messaging/model/StompMessage.java
@@ -1,0 +1,4 @@
+package com.roboter5123.backend.messaging.model;
+public interface StompMessage {
+
+}
