+++ src/main/java/com/roboter5123/backend/publicapi/model/HttpGameDTO.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.messaging.model;
+package com.roboter5123.backend.publicapi.model;
 import com.roboter5123.backend.game.api.GameMode;
 import lombok.AllArgsConstructor;
 import lombok.Data;
