+++ src/main/java/com/roboter5123/backend/publicapi/model/exception/NoSuchSessionHttpException.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.messaging.model.exception;
+package com.roboter5123.backend.publicapi.model.exception;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
