+++ src/main/java/com/roboter5123/backend/service/api/GameService.java
@@ -1,8 +1,8 @@
 package com.roboter5123.backend.service.api;
 import com.roboter5123.backend.game.api.Game;
 import com.roboter5123.backend.game.api.GameMode;
-import com.roboter5123.backend.service.model.exception.NoSuchSessionException;
-import com.roboter5123.backend.service.model.exception.TooManySessionsException;
+import com.roboter5123.backend.service.exception.NoSuchSessionException;
+import com.roboter5123.backend.service.exception.TooManySessionsException;
 
 import java.util.Map;
 import java.util.Optional;
