+++ src/main/java/com/roboter5123/backend/service/exception/NoSuchSessionException.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.service.model.exception;
+package com.roboter5123.backend.service.exception;
 public class NoSuchSessionException extends RuntimeException {
 
     public NoSuchSessionException(String message) {
