+++ src/main/java/com/roboter5123/backend/service/exception/TooManySessionsException.java
@@ -1,4 +1,4 @@
-package com.roboter5123.backend.service.model.exception;
+package com.roboter5123.backend.service.exception;
 public class TooManySessionsException extends RuntimeException {
 
 }
