+++ src/main/java/com/roboter5123/backend/service/implementation/GameServiceImpl.java
@@ -3,8 +3,8 @@ import com.roboter5123.backend.game.api.Game;
 import com.roboter5123.backend.game.api.GameMode;
 import com.roboter5123.backend.service.api.GameService;
 import com.roboter5123.backend.service.api.GameSessionManager;
-import com.roboter5123.backend.service.model.exception.NoSuchSessionException;
-import com.roboter5123.backend.service.model.exception.TooManySessionsException;
+import com.roboter5123.backend.service.exception.NoSuchSessionException;
+import com.roboter5123.backend.service.exception.TooManySessionsException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
