--- src/main/java/com/roboter5123/backend/service/model/StompMessage.java
@@ -1,4 +1,0 @@
-package com.roboter5123.backend.service.model;
-public interface StompMessage {
-
-}
