+++ src/test/java/com/roboter5123/backend/publicapi/implementation/HttpControllerImplTest.java
@@ -1,13 +1,13 @@
-package com.roboter5123.backend.messaging.implementation;
+package com.roboter5123.backend.publicapi.implementation;
 import com.roboter5123.backend.game.api.Game;
 import com.roboter5123.backend.game.api.GameMode;
-import com.roboter5123.backend.game.api.MessageBroadcaster;
+import com.roboter5123.backend.messaging.api.MessageBroadcaster;
 import com.roboter5123.backend.game.implementation.chat.ChatGame;
-import com.roboter5123.backend.messaging.api.HttpController;
-import com.roboter5123.backend.messaging.model.HttpGameDTO;
+import com.roboter5123.backend.publicapi.api.HttpController;
+import com.roboter5123.backend.publicapi.model.HttpGameDTO;
 import com.roboter5123.backend.service.api.GameService;
-import com.roboter5123.backend.service.api.StompMessageFactory;
-import com.roboter5123.backend.service.model.exception.TooManySessionsException;
+import com.roboter5123.backend.messaging.api.StompMessageFactory;
+import com.roboter5123.backend.service.exception.TooManySessionsException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
