+++ src/test/java/com/roboter5123/backend/publicapi/implementation/StompControllerImplTest.java
@@ -1,12 +1,12 @@
-package com.roboter5123.backend.messaging.implementation;
-import com.roboter5123.backend.messaging.api.StompController;
+package com.roboter5123.backend.publicapi.implementation;
+import com.roboter5123.backend.publicapi.api.StompController;
 import com.roboter5123.backend.service.api.GameService;
-import com.roboter5123.backend.game.api.MessageBroadcaster;
-import com.roboter5123.backend.service.api.StompMessageFactory;
-import com.roboter5123.backend.service.model.ErrorStompMessage;
-import com.roboter5123.backend.service.model.StateStompMessage;
-import com.roboter5123.backend.service.model.StompMessage;
-import com.roboter5123.backend.service.model.exception.NoSuchSessionException;
+import com.roboter5123.backend.messaging.api.MessageBroadcaster;
+import com.roboter5123.backend.messaging.api.StompMessageFactory;
+import com.roboter5123.backend.messaging.model.ErrorStompMessage;
+import com.roboter5123.backend.messaging.model.StateStompMessage;
+import com.roboter5123.backend.messaging.model.StompMessage;
+import com.roboter5123.backend.service.exception.NoSuchSessionException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
