+++ src/test/java/com/roboter5123/backend/service/implementation/MessageBroadcasterImplTest.java
@@ -1,7 +1,7 @@
 package com.roboter5123.backend.service.implementation;
-import com.roboter5123.backend.game.api.MessageBroadcaster;
-import com.roboter5123.backend.service.api.StompMessageFactory;
-import com.roboter5123.backend.service.model.ErrorStompMessage;
+import com.roboter5123.backend.messaging.api.MessageBroadcaster;
+import com.roboter5123.backend.messaging.api.StompMessageFactory;
+import com.roboter5123.backend.messaging.model.ErrorStompMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
