+++ src/test/java/com/roboter5123/backend/service/implementation/StompMessageFactoryImplTest.java
@@ -2,12 +2,12 @@ package com.roboter5123.backend.service.implementation;
 import com.roboter5123.backend.game.api.Command;
 import com.roboter5123.backend.game.implementation.chat.ChatMessageAction;
 import com.roboter5123.backend.game.implementation.chat.ChatState;
-import com.roboter5123.backend.service.api.StompMessageFactory;
-import com.roboter5123.backend.service.model.ErrorStompMessage;
-import com.roboter5123.backend.service.model.ServiceCommandStompMessage;
-import com.roboter5123.backend.service.model.StateStompMessage;
-import com.roboter5123.backend.service.model.StompMessage;
-import com.roboter5123.backend.service.model.ServiceCommand;
+import com.roboter5123.backend.messaging.api.StompMessageFactory;
+import com.roboter5123.backend.messaging.model.ErrorStompMessage;
+import com.roboter5123.backend.messaging.model.ServiceCommandStompMessage;
+import com.roboter5123.backend.messaging.model.StateStompMessage;
+import com.roboter5123.backend.messaging.model.StompMessage;
+import com.roboter5123.backend.messaging.model.ServiceCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
